unresolved error log and react quill bugs resolved

+++ client/src/pages/BlogDetails.jsx
@@ -1,6 +1,6 @@
 // import { blogPosts } from '../../              /data';
 import { useState, useEffect } from 'react';
-import { useParams, Navigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { useNavigate, Link } from 'react-router-dom';
 import formatDate from '../lib/formattedDate';
 import http from '../lib/http';
@@ -15,6 +15,11 @@ const BlogDetails = () => {
 
   const { user } = useAuthContext();
   const token = user?.token;
+  const role = user?.data?.user?.role;
+
+  // checks admin role for writing
+  const isAdmin = role === 'admin';
+  // console.log(role);
 
   useEffect(() => {
     async function fetchData() {
@@ -32,67 +37,67 @@ const BlogDetails = () => {
   if (!blogPosts) return null;
 
   const deletePost = async () => {
+    if (!isAdmin) return;
+
     if (!user) {
-      // setError("You must be logged in")
       return;
     }
     await http.delete(`/posts/${params.id}`, { token });
-    // deleteAlert(true);
-
     navigate('/blog');
   };
 
-  // // handles delete alert
-  // const handleDeleteAlert = () => {
-  //   setDeleteAlert(!deleteAlert);
-  // };
-
   return (
     <div className='blog-detail'>
-      {/* <div>{deleteAlert && <h1>Do you want to delete this post</h1>}</div> */}
-
       <div className='blog-detail-content'>
-        {user && (
-          <div className='edit-delete-btn'>
-            {blogPosts && (
-              <img
-                src={`http://localhost:3300/images/${blogPosts.photo}`}
-                alt={blogPosts.title}
-              />
-            )}
-
-            <div className='btns'>
-              <button className='edit'>
-                <Link to={`/blog/${params.id}/edit`}>Edit</Link>
-              </button>
-              <button
-                className='delete'
-                onClick={() => setDeleteAlert(!deleteAlert)}>
-                Delete
-              </button>
-            </div>
-            {deleteAlert ? (
-              <div>
-                <p>
-                  Are you sure you want to delete this post? if yes, click on{' '}
-                  <strong>Yes</strong> to proceed or <strong>No</strong> to
-                  cancel
-                </p>
-                <button onClick={() => deletePost()}>YES</button>
-                <button onClick={() => setDeleteAlert(!deleteAlert)}>NO</button>
+        {blogPosts && (
+          <img
+            className='blog-detail-img'
+            src={`http://localhost:3300/images/${blogPosts.photo}`}
+            alt={blogPosts.title}
+          />
+        )}
+        {/* show this only to admin */}
+        {isAdmin && (
+          <div className='edit-delete'>
+            {user && (
+              <div className='edit-delete-btn'>
+                <div className='btns'>
+                  <button className='edit'>
+                    <Link to={`/blog/${params.id}/edit`}>Edit</Link>
+                  </button>
+                  <button
+                    className='delete'
+                    onClick={() => setDeleteAlert(!deleteAlert)}>
+                    Delete
+                  </button>
+                </div>
+                {deleteAlert ? (
+                  <div>
+                    <p>
+                      Are you sure you want to delete this post? if yes, click
+                      on <strong>Yes</strong> to proceed or <strong>No</strong>{' '}
+                      to cancel
+                    </p>
+                    <button onClick={() => deletePost()}>YES</button>
+                    <button onClick={() => setDeleteAlert(!deleteAlert)}>
+                      NO
+                    </button>
+                  </div>
+                ) : (
+                  ''
+                )}
               </div>
-            ) : (
-              ''
             )}
           </div>
         )}
-
         <h1>{blogPosts.title}</h1>
         <div className='blog-content-author-date'>
           <span>{formatDate(blogPosts.date)}</span>
           <span>By: {blogPosts.author}</span>
         </div>
-        <p>{blogPosts.body.split('/n').join(' ')}</p>
+        {/* <p>{blogPosts.body.split('/n').join(``)}</p> */}
+        <p dangerouslySetInnerHTML={{ __html: blogPosts.body }}></p>
+        {/* <p dangerouslySetInnerHTML={{ __html: body.substring(0, 100) }} /> */}
 
         <Link to='/blog' style={{ textDecoration: 'none' }}>
           &#8592; Go back
